Move new geometry tests into resstockarguments and refactor config steps.
