+++ .github/workflows/config.yml
@@ -78,7 +78,7 @@ jobs:
           path: coverage
           name: coverage
 
-  build-existing-model-tests:
+  measure-tests:
     runs-on: ubuntu-latest
     needs: [unit-tests]
     container:
@@ -94,7 +94,7 @@ jobs:
 
       - name: Run BuildExistingModel measure tests
         run: |
-          bundle exec rake unit_tests:build_existing_model_tests
+          bundle exec rake workflow:measure_tests
 
   analysis-tests:
     runs-on: ubuntu-latest
