+++ measures/BuildExistingModel/measure.xml
@@ -3,8 +3,8 @@
   <schema_version>3.1</schema_version>
   <name>build_existing_model</name>
   <uid>dedf59bb-3b88-4f16-8755-2c1ff5519cbf</uid>
-  <version_id>7f2fb3c6-7da6-4014-8dde-8ee477638c86</version_id>
-  <version_modified>2024-06-18T04:19:07Z</version_modified>
+  <version_id>e617045c-db32-4120-8c7c-c73a1abd482b</version_id>
+  <version_modified>2024-06-18T15:54:30Z</version_modified>
   <xml_checksum>2C38F48B</xml_checksum>
   <class_name>BuildExistingModel</class_name>
   <display_name>Build Existing Model</display_name>
@@ -358,12 +358,6 @@
       <filetype>rb</filetype>
       <usage_type>script</usage_type>
       <checksum>4CFF6DBA</checksum>
-    </file>
-    <file>
-      <filename>build_existing_model_test.rb</filename>
-      <filetype>rb</filetype>
-      <usage_type>test</usage_type>
-      <checksum>CF2818E1</checksum>
     </file>
   </files>
 </measure>
