+++ measures/ResStockArguments/measure.xml
@@ -3,8 +3,8 @@
   <schema_version>3.1</schema_version>
   <name>res_stock_arguments</name>
   <uid>c984bb9e-4ac4-4930-a399-9d23f8f6936a</uid>
-  <version_id>5ad2e474-6d5a-404f-9976-6d835d0614d3</version_id>
-  <version_modified>2024-05-20T19:45:30Z</version_modified>
+  <version_id>4527e5bd-c704-43a9-893f-784bedcf1d90</version_id>
+  <version_modified>2024-06-18T16:05:37Z</version_modified>
   <xml_checksum>2C38F48B</xml_checksum>
   <class_name>ResStockArguments</class_name>
   <display_name>ResStock Arguments</display_name>
@@ -7509,7 +7509,7 @@
       <filename>resstock_arguments_test.rb</filename>
       <filetype>rb</filetype>
       <usage_type>test</usage_type>
-      <checksum>6DD62440</checksum>
+      <checksum>033B39BC</checksum>
     </file>
   </files>
 </measure>
